Support Long type config values
### What changes are proposed in this pull request?

Revert getLong API

### Why are the changes needed?
There are some offline discussions, it is believed that some
`propertyKeys` may require a value of type Long, so the revert getLong
API is required.

Use case: add a propertykey to limit the capacity of inode, it should be
defined as long

### Does this PR introduce any user facing changes?
No

pr-link: Alluxio/alluxio#15444

+++ core/server/common/src/main/java/alluxio/master/journal/raft/RaftJournalSystem.java
@@ -378,8 +378,8 @@ public class RaftJournalSystem extends AbstractJournalSystem {
     // snapshot interval
     RaftServerConfigKeys.Snapshot.setAutoTriggerEnabled(
         properties, true);
-    int snapshotAutoTriggerThreshold =
-        Configuration.getInt(PropertyKey.MASTER_JOURNAL_CHECKPOINT_PERIOD_ENTRIES);
+    long snapshotAutoTriggerThreshold =
+        Configuration.getLong(PropertyKey.MASTER_JOURNAL_CHECKPOINT_PERIOD_ENTRIES);
     RaftServerConfigKeys.Snapshot.setAutoTriggerThreshold(properties,
         snapshotAutoTriggerThreshold);
     // minimum distance between two snapshots in number of log entries
