+++ job/server/src/main/java/alluxio/master/job/JobMaster.java
@@ -154,9 +154,9 @@ public class JobMaster extends AbstractMaster implements NoopJournaled {
     mWorkflowTracker = new WorkflowTracker(this);
 
     mPlanTracker = new PlanTracker(
-        Configuration.getInt(PropertyKey.JOB_MASTER_JOB_CAPACITY),
+        Configuration.getLong(PropertyKey.JOB_MASTER_JOB_CAPACITY),
         Configuration.getMs(PropertyKey.JOB_MASTER_FINISHED_JOB_RETENTION_TIME),
-        Configuration.getInt(PropertyKey.JOB_MASTER_FINISHED_JOB_PURGE_COUNT),
+        Configuration.getLong(PropertyKey.JOB_MASTER_FINISHED_JOB_PURGE_COUNT),
         mWorkflowTracker);
 
     mWorkerHealth = new ConcurrentHashMap<>();
